+++ pxr/usd/lib/sdf/specType.cpp
@@ -52,7 +52,7 @@ PXR_NAMESPACE_OPEN_SCOPE
 static inline size_t
 _GetBitmaskForSpecType(SdfSpecType specType)
 {
-    return (1 << specType);
+    return (size_t(1) << specType);
 }
 
 struct Sdf_SpecTypeInfo
@@ -249,7 +249,7 @@ _CanCast(SdfSpecType fromType, const TfType& toType)
     }
 
     const size_t allowedBitmask = 
-        TfMapLookupByValue(specTypeInfo.specTypeToBitmask, toType, 0);
+        TfMapLookupByValue(specTypeInfo.specTypeToBitmask, toType, size_t(0));
     return allowedBitmask & _GetBitmaskForSpecType(fromType);
 }
